requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
python-multipart==0.0.6
aiosqlite==0.19.0
